% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_calinski_harabasz}
\alias{calc_calinski_harabasz}
\title{Calculate Calinski-Harabasz Index (CH)}
\usage{
calc_calinski_harabasz(data, cluster_labels)
}
\arguments{
\item{data}{Matrix with features in rows, cells in columns.}

\item{cluster_labels}{Cluster assignments.}
}
\value{
Calinski-Harabasz index.
}
\description{
Calculate Calinski-Harabasz Index (CH)
}
\examples{
data <- matrix(stats::rnorm(200), 20, 10)
cl <- factor(rep(c("A", "B"), each = 10))
dist_mat <- stats::dist(data)
calc_calinski_harabasz(data, cl)
}
