% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_ks}
\alias{calc_ks}
\title{Calculate Kolmogorov-Smirnov Distance (1D KS)}
\usage{
calc_ks(ref, sim)
}
\arguments{
\item{ref}{Numeric vector of reference values.}

\item{sim}{Numeric vector of simulated values.}
}
\value{
Maximum vertical distance between empirical CDFs (between 0 and 1).
}
\description{
Calculate Kolmogorov-Smirnov Distance (1D KS)
}
\examples{
ref <- stats::rnorm(50)
sim <- stats::rnorm(50)
calc_ks(ref, sim)
}
