% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_mae}
\alias{calc_mae}
\title{Calculate Mean Absolute Error (MAE)}
\usage{
calc_mae(ref, sim, align = TRUE)
}
\arguments{
\item{ref}{Numeric vector of reference values.}

\item{sim}{Numeric vector of simulated values.}

\item{align}{Logical, whether to sort and quantile-align vectors. Default TRUE.}
}
\value{
Mean absolute difference.
}
\description{
Calculate Mean Absolute Error (MAE)
}
\examples{
ref <- stats::rnorm(50)
sim <- stats::rnorm(50)
calc_mae(ref, sim)
}
