% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_metrics_de_signal.R
\name{calc_signal_de}
\alias{calc_signal_de}
\title{Detect Differential Expression (DE) Genes (Mean Shift via Limma or T-Test)}
\usage{
calc_signal_de(exprs_mat, cell_types, p_sig = 0.05)
}
\arguments{
\item{exprs_mat}{Log-normalized matrix (genes x cells).}

\item{cell_types}{Factor or binary vector of 2 cell types.}

\item{p_sig}{Significance threshold (default 0.05).}
}
\value{
Vector of adjusted p-values.
}
\description{
Detect Differential Expression (DE) Genes (Mean Shift via Limma or T-Test)
}
\examples{
data(example_scrna, package = "scSimEval")
calc_signal_de(example_scrna$ref, example_scrna$cell_types)
}
