% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_metrics_de_signal.R
\name{calc_signal_dv}
\alias{calc_signal_dv}
\title{Detect Differential Variability (DV) Genes (Bartlett Test)}
\usage{
calc_signal_dv(exprs_mat, cell_types)
}
\arguments{
\item{exprs_mat}{Log-normalized matrix (genes x cells).}

\item{cell_types}{Factor of 2 cell types.}
}
\value{
Vector of adjusted p-values.
}
\description{
Detect Differential Variability (DV) Genes (Bartlett Test)
}
\examples{
data(example_scrna, package = "scSimEval")
calc_signal_dv(example_scrna$ref, example_scrna$cell_types)
}
