% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_metrics_trajectory.R
\name{calc_tree_height_discrepancy}
\alias{calc_tree_height_discrepancy}
\title{Calculate Lineage Tree Branch Height Discrepancy (RMSE)}
\usage{
calc_tree_height_discrepancy(
  ref_tree,
  sim_tree,
  cell_types_ref = NULL,
  cell_types_sim = NULL
)
}
\arguments{
\item{ref_tree}{Hierarchical clustering (\code{hclust}) object from reference data,
OR reference count matrix.}

\item{sim_tree}{Hierarchical clustering (\code{hclust}) object from simulated data,
OR simulated count matrix.}

\item{cell_types_ref}{Optional cell type labels for reference (if count matrix provided).}

\item{cell_types_sim}{Optional cell type labels for simulation (if count matrix provided).}
}
\value{
Root mean squared error between branch heights.
}
\description{
Computes the Root Mean Squared Error (RMSE) between sorted branch heights of
hierarchical lineage trees. Accepts either precomputed \code{hclust} objects
OR scRNA-seq count matrices + \code{cell_types} (which will be automatically inferred).
}
\examples{
data(example_scrna, package = "scSimEval")
calc_tree_height_discrepancy(example_scrna$ref, example_scrna$sim)
}
