% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_visualizations.R
\name{.ingest_bubble_data}
\alias{.ingest_bubble_data}
\title{Ingest bubble matrix input into a tidy data.frame}
\usage{
.ingest_bubble_data(data)
}
\description{
Ingest bubble matrix input into a tidy data.frame
}
\keyword{internal}
