% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_extract_properties.R
\name{evaluate_accessibility_sparsity_curve}
\alias{evaluate_accessibility_sparsity_curve}
\title{Evaluate Accessibility-Sparsity Curve Concordance (simATAC)}
\usage{
evaluate_accessibility_sparsity_curve(ref_data, sim_data, poly_degree = 2)
}
\arguments{
\item{ref_data}{Reference count matrix (peaks x cells).}

\item{sim_data}{Simulated count matrix (peaks x cells).}

\item{poly_degree}{Degree of polynomial (default: 2).}
}
\value{
A list of reference and simulated curve parameters, absolute discrepancies,
  and curve prediction RMSE.
}
\description{
Compares the non-linear relationship between peak mean accessibility and non-zero
proportion (NZP) between reference and simulated scATAC-seq datasets.
}
\examples{
data(example_scrna, package = "scSimEval")
evaluate_accessibility_sparsity_curve(example_scrna$ref, example_scrna$sim)
}
