% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_metrics_trajectory.R
\name{infer_scrna_lineage_tree}
\alias{infer_scrna_lineage_tree}
\title{Automatically Infer Lineage Tree (hclust) from scRNA-seq Counts & Cell Types}
\usage{
infer_scrna_lineage_tree(data, cell_types, method = "ward.D2")
}
\arguments{
\item{data}{Expression or count matrix (genes x cells).}

\item{cell_types}{Factor or vector of cell-type annotations for each cell.}

\item{method}{Linkage method for hierarchical clustering (default "ward.D2").}
}
\value{
An \code{hclust} tree object representing the cell-type lineage tree.
}
\description{
Builds a hierarchical clustering lineage tree between cell-type centroids
from scRNA-seq expression data.
}
\examples{
data(example_scrna, package = "scSimEval")
infer_scrna_lineage_tree(example_scrna$ref, example_scrna$cell_types)
}
