% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_metrics_trajectory.R
\name{infer_scrna_pseudotime}
\alias{infer_scrna_pseudotime}
\title{Automatically Infer Pseudotime Trajectory from scRNA-seq Counts}
\usage{
infer_scrna_pseudotime(data, n_top = 500)
}
\arguments{
\item{data}{Expression or count matrix (genes x cells).}

\item{n_top}{Number of top variable genes to use (default 500).}
}
\value{
Numeric vector of pseudotime values in [0, 1] for each cell.
}
\description{
Estimates cell differentiation pseudotime directly from single-cell expression
counts using diffusion/principal curve projection along the first principal component
of top highly variable genes.
}
\examples{
data(example_scrna, package = "scSimEval")
infer_scrna_pseudotime(example_scrna$ref)
}
