% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_all_bivariate_metrics}
\alias{calc_all_bivariate_metrics}
\title{Compute All Bivariate (2D) Distance & Accuracy Metrics for Joint Distributions}
\usage{
calc_all_bivariate_metrics(ref_mat, sim_mat, metric_prefix = "", threads = 1)
}
\arguments{
\item{ref_mat}{2-column matrix for reference.}

\item{sim_mat}{2-column matrix for simulation.}

\item{metric_prefix}{Optional prefix string for metric names.}

\item{threads}{CPU threads.}
}
\value{
Named list of bivariate metrics.
}
\description{
Compute All Bivariate (2D) Distance & Accuracy Metrics for Joint Distributions
}
