% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_bhattacharyya}
\alias{calc_bhattacharyya}
\title{Calculate Bhattacharyya Distance (BH)}
\usage{
calc_bhattacharyya(ref, sim, align = TRUE)
}
\arguments{
\item{ref}{Numeric vector of reference values.}

\item{sim}{Numeric vector of simulated values.}

\item{align}{Logical, whether to align lengths. Default TRUE.}
}
\value{
Statistical divergence between discrete probability measures.
}
\description{
Calculate Bhattacharyya Distance (BH)
}
