% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_calinski_harabasz}
\alias{calc_calinski_harabasz}
\title{Calculate Calinski-Harabasz Index (CH)}
\usage{
calc_calinski_harabasz(data, cluster_labels)
}
\arguments{
\item{data}{Matrix with features in rows, cells in columns.}

\item{cluster_labels}{Cluster assignments.}
}
\value{
Calinski-Harabasz index.
}
\description{
Calculate Calinski-Harabasz Index (CH)
}
