% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_completeness}
\alias{calc_completeness}
\title{Calculate Completeness Score}
\usage{
calc_completeness(pred, truth)
}
\arguments{
\item{pred}{Vector of predicted cluster labels.}

\item{truth}{Vector of ground truth cell type labels.}
}
\value{
Completeness score between 0 and 1.
}
\description{
Calculate Completeness Score
}
