% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_connectivity}
\alias{calc_connectivity}
\title{Calculate Cluster Connectivity}
\usage{
calc_connectivity(dist_mat, cluster_labels)
}
\arguments{
\item{dist_mat}{Distance matrix or numeric matrix.}

\item{cluster_labels}{Cluster assignments.}
}
\value{
Connectivity metric.
}
\description{
Calculate Cluster Connectivity
}
