% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_davies_bouldin}
\alias{calc_davies_bouldin}
\title{Calculate Davies-Bouldin Index (DB)}
\usage{
calc_davies_bouldin(data, cluster_labels)
}
\arguments{
\item{data}{Matrix with features in rows, cells in columns.}

\item{cluster_labels}{Cluster assignments.}
}
\value{
Davies-Bouldin index (lower is better).
}
\description{
Calculate Davies-Bouldin Index (DB)
}
\examples{
data <- matrix(stats::rnorm(200), 20, 10)
cl <- factor(rep(c("A", "B"), each = 10))
dist_mat <- stats::dist(data)
calc_davies_bouldin(data, cl)
}
