% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_dunn}
\alias{calc_dunn}
\title{Calculate Dunn Index}
\usage{
calc_dunn(dist_mat, cluster_labels)
}
\arguments{
\item{dist_mat}{Distance matrix or numeric matrix.}

\item{cluster_labels}{Cluster assignments.}
}
\value{
Dunn index (ratio of smallest inter-cluster to largest intra-cluster distance).
}
\description{
Calculate Dunn Index
}
