% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_fasano_franceschini}
\alias{calc_fasano_franceschini}
\title{Calculate Fasano-Franceschini 2D Kolmogorov-Smirnov Test Statistic}
\usage{
calc_fasano_franceschini(ref_mat, sim_mat, threads = 1)
}
\arguments{
\item{ref_mat}{2-column numeric matrix for reference.}

\item{sim_mat}{2-column numeric matrix for simulation.}

\item{threads}{CPU threads. Default 1.}
}
\value{
Estimated 2D KS statistic.
}
\description{
Integrated from simpipe.
}
\examples{
ref <- matrix(stats::rnorm(100), 50, 2)
sim <- matrix(stats::rnorm(100), 50, 2)
calc_fasano_franceschini(ref, sim)
}
