% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_fmi}
\alias{calc_fmi}
\title{Calculate Fowlkes-Mallows Index (FMI)}
\usage{
calc_fmi(pred, truth)
}
\arguments{
\item{pred}{Vector of predicted cluster labels.}

\item{truth}{Vector of ground truth cell type labels.}
}
\value{
FMI score between 0 and 1.
}
\description{
Integrated from scCluBench (Xu et al., AAAI 2026). Computes the geometric mean
of pairwise cluster precision and recall.
}
