% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_kde_test}
\alias{calc_kde_test}
\title{Calculate 2D Bivariate Kernel Density Estimation (KDE) Test Statistic
Integrated from simpipe & SimBench.}
\usage{
calc_kde_test(ref_mat, sim_mat)
}
\arguments{
\item{ref_mat}{2-column numeric matrix for reference.}

\item{sim_mat}{2-column numeric matrix for simulation.}
}
\value{
z-statistic testing discrepancy between bivariate kernel density estimates.
}
\description{
Calculate 2D Bivariate Kernel Density Estimation (KDE) Test Statistic
Integrated from simpipe & SimBench.
}
\examples{
ref <- matrix(stats::rnorm(100), 50, 2)
sim <- matrix(stats::rnorm(100), 50, 2)
calc_kde_test(ref, sim)
}
