% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_mad}
\alias{calc_mad}
\title{Calculate Median Absolute Deviation (MAD)}
\usage{
calc_mad(ref, sim, align = TRUE)
}
\arguments{
\item{ref}{Numeric vector of reference distribution.}

\item{sim}{Numeric vector of simulated distribution.}

\item{align}{Logical, whether to sort and quantile-align vectors. Default TRUE.}
}
\value{
Numeric MAD value.
}
\description{
Calculate Median Absolute Deviation (MAD)
}
