% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{calc_outlier_proportion}
\alias{calc_outlier_proportion}
\title{Calculate Proportion of Outliers in a Numeric Vector}
\usage{
calc_outlier_proportion(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Outlier proportion (between 0 and 1).
}
\description{
Uses the standard Tukey 1.5 * IQR criterion on quantiles, as used in simpipe.
}
