% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_overlap}
\alias{calc_overlap}
\title{Calculate Distribution Overlapping Index (OV)}
\usage{
calc_overlap(ref, sim)
}
\arguments{
\item{ref}{Numeric vector of reference values.}

\item{sim}{Numeric vector of simulated values.}
}
\value{
Area of overlap under probability densities (0 to 1).
}
\description{
Calculate Distribution Overlapping Index (OV)
}
