% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_peacock_2d}
\alias{calc_peacock_2d}
\title{Calculate Peacock 2D Kolmogorov-Smirnov Test Statistic
Integrated from HelenaLC/simulation-comparison.}
\usage{
calc_peacock_2d(ref_mat, sim_mat)
}
\arguments{
\item{ref_mat}{2-column numeric matrix for reference.}

\item{sim_mat}{2-column numeric matrix for simulation.}
}
\value{
Peacock test statistic.
}
\description{
Calculate Peacock 2D Kolmogorov-Smirnov Test Statistic
Integrated from HelenaLC/simulation-comparison.
}
\examples{
ref <- matrix(stats::rnorm(100), 50, 2)
sim <- matrix(stats::rnorm(100), 50, 2)
calc_peacock_2d(ref, sim)
}
