% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics_distribution.R
\name{calc_rmse}
\alias{calc_rmse}
\title{Calculate Root Mean Squared Error (RMSE)}
\usage{
calc_rmse(ref, sim, align = TRUE)
}
\arguments{
\item{ref}{Numeric vector of reference values.}

\item{sim}{Numeric vector of simulated values.}

\item{align}{Logical, whether to sort and quantile-align vectors. Default TRUE.}
}
\value{
Quadratic error penalizing large discrepancies.
}
\description{
Calculate Root Mean Squared Error (RMSE)
}
