% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_metrics_de_signal.R
\name{calc_signal_bd}
\alias{calc_signal_bd}
\title{Detect Bimodal Distribution (BD) Genes (Bimodal Separation Index)}
\usage{
calc_signal_bd(exprs_mat, cell_types)
}
\arguments{
\item{exprs_mat}{Matrix (genes x cells).}

\item{cell_types}{Factor of 2 cell types.}
}
\value{
Numeric vector of Bimodality Index values.
}
\description{
Detect Bimodal Distribution (BD) Genes (Bimodal Separation Index)
}
