% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_metrics_de_signal.R
\name{calc_signal_dd}
\alias{calc_signal_dd}
\title{Detect Differential Distribution (DD) Genes (Kolmogorov-Smirnov Test)}
\usage{
calc_signal_dd(exprs_mat, cell_types)
}
\arguments{
\item{exprs_mat}{Log-normalized matrix (genes x cells).}

\item{cell_types}{Factor of 2 cell types.}
}
\value{
Vector of adjusted p-values.
}
\description{
Detect Differential Distribution (DD) Genes (Kolmogorov-Smirnov Test)
}
