% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_metrics_de_signal.R
\name{calc_signal_dp}
\alias{calc_signal_dp}
\title{Detect Differential Proportion / Zero-Inflation (DP) Genes (Chisq Test)}
\usage{
calc_signal_dp(exprs_mat, cell_types, threshold = 0)
}
\arguments{
\item{exprs_mat}{Count or expression matrix (genes x cells).}

\item{cell_types}{Factor of 2 cell types.}

\item{threshold}{Value threshold for zero/detection status (default 0).}
}
\value{
Vector of adjusted p-values.
}
\description{
Detect Differential Proportion / Zero-Inflation (DP) Genes (Chisq Test)
}
