% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_silhouette}
\alias{calc_silhouette}
\title{Calculate Average Silhouette Width (ASW)}
\usage{
calc_silhouette(dist_mat, cluster_labels)
}
\arguments{
\item{dist_mat}{Distance matrix among cells or numeric matrix (features x cells).}

\item{cluster_labels}{Vector of cluster or cell-type labels.}
}
\value{
Mean silhouette width.
}
\description{
Calculate Average Silhouette Width (ASW)
}
