% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metrics_clustering.R
\name{calc_v_measure}
\alias{calc_v_measure}
\title{Calculate V-Measure}
\usage{
calc_v_measure(pred, truth, beta = 1)
}
\arguments{
\item{pred}{Vector of predicted cluster labels.}

\item{truth}{Vector of ground truth cell type labels.}

\item{beta}{Weight parameter (default 1).}
}
\value{
V-measure score between 0 and 1.
}
\description{
Calculate V-Measure
}
